+++ src/action-types.js
@@ -1,0 +1,6 @@
+const ActionType = {
+  SELECT_FILTER: `SELECT_FILTER`,
+  GET_FILTERED_LIST: `GET_FILTERED_LIST`,
+};
+
+export default ActionType;
