+++ src/initial-state.js
@@ -1,0 +1,9 @@
+import films from "./mocks/films.js";
+
+const initialState = {
+  genre: `All genres`,
+  filteredList: [...films],
+  films,
+};
+
+export default initialState;
